+++ packages/sandbox/src/lambda_function_log_streamer.ts
@@ -1,5 +1,9 @@
 import { LogLevel, Printer } from '@aws-amplify/cli-core';
-import { BackendOutputClient } from '@aws-amplify/deployed-backend-client';
+import {
+  BackendOutputClient,
+  BackendOutputClientError,
+  BackendOutputClientErrorType,
+} from '@aws-amplify/deployed-backend-client';
 import { TagName } from '@aws-amplify/platform-core';
 import { BackendIdentifier, BackendOutput } from '@aws-amplify/plugin-types';
 
@@ -37,8 +41,21 @@ export class LambdaFunctionLogStreamer {
       return;
     }
 
-    const backendOutput: BackendOutput =
-      await this.backendOutputClient.getOutput(sandboxBackendId);
+    let backendOutput: BackendOutput = {};
+    try {
+      backendOutput = await this.backendOutputClient.getOutput(
+        sandboxBackendId
+      );
+    } catch (error) {
+      // If stack does not exist, we do not want to go further to start streaming logs
+      if (
+        BackendOutputClientError.isBackendOutputClientError(error) &&
+        error.code === BackendOutputClientErrorType.NO_STACK_FOUND
+      ) {
+        this.enabled = false;
+        return;
+      }
+    }
 
     const definedFunctionsPayload =
       backendOutput['AWS::Amplify::Function']?.payload.definedFunctions;